# doc/api.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    {
      post:
        {
          operationId: AuthController_signIn,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successfully signed in',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            { $ref: '#/components/schemas/GetTokensResponse' },
                        },
                    },
                },
              '401': { description: 'Incorrect login or password' },
            },
          tags: [authentication],
        },
    }
  /auth/signup:
    {
      post:
        {
          operationId: AuthController_signUp,
          parameters: [],
          requestBody:
            {
              required: true,
              content:
                {
                  application/json:
                    { schema: { $ref: '#/components/schemas/CreateUserDto' } },
                },
            },
          responses:
            {
              '201':
                {
                  description: 'Successfully signed up. To continue, please, confirm your email',
                  content:
                    {
                      application/json:
                        { schema: { $ref: '#/components/schemas/User' } },
                    },
                },
              '400': { description: 'Invalid dto' },
              '409': { description: 'Login or email already in use' },
            },
          tags: [authentication],
        },
    }
  /auth/refresh:
    {
      post:
        {
          operationId: AuthController_refresh,
          parameters: [],
          responses:
            {
              '200':
                {
                  description: 'Successfully refreshed',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            { $ref: '#/components/schemas/GetTokensResponse' },
                        },
                    },
                },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [authentication],
          security: [{ bearer: [] }],
        },
    }
  /auth/confirm:
    {
      get:
        {
          operationId: AuthController_confirmUserEmail,
          parameters:
            [
              {
                name: token,
                required: true,
                in: query,
                description: 'Email confirmation token',
                schema: {},
              },
            ],
          responses:
            {
              '200':
                {
                  description: 'Successfully confirmed',
                  content:
                    {
                      application/json:
                        {
                          schema:
                            { $ref: '#/components/schemas/GetTokensResponse' },
                        },
                    },
                },
              '401': { description: Unauthorized },
              '403': { description: 'Operation forbidden' },
            },
          tags: [authentication],
        },
    }
info:
  title: 'Home Library'
  description: 'Home music library service'
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes: { bearer: { scheme: bearer, bearerFormat: JWT, type: http } }
  schemas:
...
      User:
        {
          type: object,
          properties:
            {
              id: { type: string },
              login: { type: string },
              email: { type: string },
              emailIsConfirmed: { type: boolean },
              version: { type: number },
              createdAt: { type: number },
              updatedAt: { type: number },
              role: { type: string },
            },
          required:
            [
              id,
              login,
              email,
              emailIsConfirmed,
              version,
              createdAt,
              updatedAt,
              role,
            ],
        },
...
      GetTokensResponse:
        {
          type: object,
          properties:
            { accessToken: { type: string }, refreshToken: { type: string } },
          required: [accessToken, refreshToken],
        },
      CreateUserDto:
        {
          type: object,
          properties:
            {
              login: { type: string },
              email: { type: string },
              password: { type: string },
            },
          required: [login, email, password],
        },
...